requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
Pillow>=5.0.0
factory_boy>=3.1.0
djangorestframework-jwt>=1.11.0
drf-yasg>=1.17.1